Refactor LinkedIn agent
