+++ ice_breaker/requirements.txt
@@ -1,2 +1,3 @@
 langchain==0.0.138
 openai==0.27.4
+google-search-results==2.4.2
